requirements.txt
asgiref==3.3.1
astroid==2.4.2
colorama==0.4.4
Django==3.1.5
django-appconf==1.0.4
django-user-accounts==3.0.2
djangorestframework==3.12.2
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
pytz==2020.5
six==1.15.0
SQLAlchemy==1.3.22
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
